+++ src/ethereum/EthereumEngine.ts
@@ -79,7 +79,11 @@ import {
   SafeEthWalletInfo,
   TxRpcParams
 } from './ethereumTypes'
-import { calcL1RollupFees, calcMiningFees } from './fees/ethMiningFees'
+import {
+  calcL1RollupFees,
+  calcMiningFees,
+  getFeeParamsByTransactionType
+} from './fees/ethMiningFees'
 import {
   FeeProviderFunction,
   FeeProviders,
@@ -624,7 +628,7 @@ export class EthereumEngine extends CurrencyEngine<
     // eslint-disable-next-line @typescript-eslint/strict-boolean-expressions
     if (!supportsEIP1559) return
 
-    const { baseFeePerGas } = await this.ethNetwork.getBaseFeePerGas()
+    const baseFeePerGas = await this.ethNetwork.getBaseFeePerGas()
     if (baseFeePerGas == null) return
     const baseFeePerGasDecimal = hexToDecimal(baseFeePerGas)
 
@@ -1177,14 +1181,24 @@ export class EthereumEngine extends CurrencyEngine<
     const { chainParams } = this.networkInfo
     const common = Common.custom(chainParams)
 
+    // Translate legacy transaction types to EIP-1559 transaction type
+    const txType = this.networkInfo.supportsEIP1559 === true ? 2 : 1
+    // Translate legacy transaction types gas params to to EIP-1559 params
+    const gasFeeParams = await getFeeParamsByTransactionType(
+      txType,
+      gasPriceHex,
+      this.ethNetwork.getBaseFeePerGas
+    )
+
     // Transaction Parameters
     const txParams = {
       nonce: nonceHex,
-      gasPrice: gasPriceHex,
+      ...gasFeeParams,
       gasLimit: gasLimitHex,
       to: otherParams.to[0],
       value: txValue,
-      data
+      data,
+      type: txType
     }
 
     const privKey = Buffer.from(ethereumPrivateKeys.privateKey, 'hex')